Fixes webpack config
