+++ webpack.config.js
@@ -1,0 +1,21 @@
+const path = require('path');
+const fs = require('fs');
+const webpack = require('webpack');
+
+var nodeModules = {};
+fs.readdirSync('node_modules')
+  .filter(function(x) {
+    return ['.bin'].indexOf(x) === -1;
+  })
+  .forEach(function(mod) {
+    nodeModules[mod] = 'commonjs ' + mod;
+  });
+
+module.exports = {
+  entry: './src/server/index.js',
+  target: node,
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'bundle.js'
+  }
+};
